@@ -1,6 +1,6 @@
 module github.com/redhat-developer/app-services-cli
 
-go 1.16
+go 1.18
 
 require (
 	github.com/AlecAivazis/survey/v2 v2.3.5
