@@ -3,7 +3,6 @@ package servicecontext
 import (
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
 	"os"
 	"path/filepath"
 )
@@ -38,7 +37,7 @@ func (c *File) Load() (*Context, error) {
 		return nil, fmt.Errorf(errorFormat, "unable to check if context file exists", err)
 	}
 	// #nosec G304
-	data, err := ioutil.ReadFile(file)
+	data, err := os.ReadFile(file)
 	if err != nil {
 		return nil, fmt.Errorf(errorFormat, "unable to read context file", err)
 	}
@@ -70,7 +69,7 @@ func (c *File) Save(cfg *Context) error {
 			return err
 		}
 	}
-	err = ioutil.WriteFile(file, data, 0o600)
+	err = os.WriteFile(file, data, 0o600)
 	if err != nil {
 		return fmt.Errorf(errorFormat, "unable to save context", err)
 	}
